+++ tests/unit-global/esnext.iterator.zip-keyed.js
@@ -17,6 +17,7 @@ QUnit.test('Iterator.zipKeyed', assert => {
   assert.nonEnumerable(Iterator, 'zipKeyed');
 
   let result = zipKeyed({ a: [0, 1, 2], b: [3, 4, 5], c: [7, 8, 9] });
+  assert.true(result instanceof Iterator, 'Iterator instance');
   assert.deepEqual(from(result), [{ a: 0, b: 3, c: 7 }, { a: 1, b: 4, c: 8 }, { a: 2, b: 5, c: 9 }]);
   result = zipKeyed({ a: [0, 1, 2], b: [3, 4, 5, 6], c: [7, 8, 9] });
   assert.deepEqual(from(result), [{ a: 0, b: 3, c: 7 }, { a: 1, b: 4, c: 8 }, { a: 2, b: 5, c: 9 }]);